refactor: 불필요한 useCallback 제거 및 이름 변경
