+++ src/app/lib/hooks/post/useSaveDraft.ts
@@ -1,27 +1,27 @@
-import { useCallback } from 'react';
 import { toast } from 'react-toastify';
 import { useCurrentEditor } from '@tiptap/react';
 import { usePostContent } from '@/app/lib/contexts/post/PostContentContext';
-import { createDraft, editDraft } from '@/app/lib/apis/post/submitDraft';
+import { createDraft, editDraft } from '@/app/lib/apis/post/saveDraft';
 
-export const useSubmitDraft = (close: () => void) => {
-  const { boardId, setBoardId, title, tagList, boardHeadImageUrl } = usePostContent();
+export const useSaveDraft = (close: () => void) => {
+  const { boardId, setBoardId, title, tagList, boardHeadImageUrl } =
+    usePostContent();
   const { editor } = useCurrentEditor();
 
-  const submitDraft = useCallback(async () => {
+  const handleSuccess = () => {
+    close();
+    toast.success('작성 중인 글이 저장되었습니다.');
+  };
+
+  const handleError = () => toast.error('작성 글 저장에 실패했습니다.');
+
+  const saveDraft = async () => {
     if (!editor) return;
 
-    if(title.length === 0){
+    if (title.length === 0) {
       toast.warn('제목을 작성해주세요.');
       return;
     }
-
-    const handleSuccess = () => {
-      close();
-      toast.success('작성 중인 글이 저장되었습니다.');
-    };
-
-    const handleError = () => toast.error('작성 글 저장에 실패했습니다.');
 
     if (boardId) {
       await editDraft(
@@ -51,7 +51,7 @@ export const useSubmitDraft = (close: () => void) => {
         setBoardId(newBoardId);
       }
     }
-  }, [boardId, editor, title, tagList, boardHeadImageUrl, close, setBoardId]);
+  };
 
-  return { submitDraft };
+  return { saveDraft };
 };
